requirements.txt
GitPython==2.1.10
termcolor==1.1.0